requirements.txt
pyautogui
pydirectinput
pyperclip
pytesseract
opencv-python
numpy
pywin32
PyYAML
colorlog
